replaced unsigned short -> uint8_t as per issue #108

+++ main/utils.h
@@ -28,7 +28,7 @@ struct bmpReading //all data from an altimeter reading
 {
   double altitude;
   uint32_t time;
-  unsigned short state; //used for tracking flightPhase (0-3 representing ONPAD - POST_FLIGHT)
+  uint8_t state; //used for tracking flightPhase (0-3 representing ONPAD - POST_FLIGHT)
 };
 
 struct imuReading //all data from an IMU reading (accelerometer and gyroscope)
